apps/api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

sqlalchemy>=2.0.25
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
jsonschema>=4.21.0

python-jose[cryptography]>=3.3.0
bcrypt>=4.0.1

redis>=5.0.1
openai>=1.10.0
tenacity>=8.2.3

email-validator>=2.1.0
python-dotenv>=1.0.0
sentry-sdk[fastapi]>=1.40.0
prometheus-client>=0.19.0
